backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv>=1.0.0
httpx==0.26.0
google-generativeai>=0.3.0
pydantic==2.6.1
moviepy>=1.0.3
yt-dlp>=2023.12.30
requests>=2.31.0
Pillow[webp,tiff,icns]>=10.1.0
numpy>=1.26.0
ffmpeg-python>=0.2.0
pytubefix
